+++ src/management/src/management_frontend_tests.cpp
@@ -50,6 +50,7 @@ TEST_F(ManagementFrontendTest, TestAddApplicationCorrect) {
     "candidate_model_names": ["m", "image_model"],
     "input_type": "integers",
     "selection_policy": "sample_policy",
+    "default_output": "4.3",
     "latency_slo_micros": 10000
   }
   )";
@@ -59,7 +60,7 @@ TEST_F(ManagementFrontendTest, TestAddApplicationCorrect) {
   // The application table has 5 fields, so we expect to get back a map with 5
   // entries in it (see add_application() in redis.cpp for details on what the
   // fields are).
-  ASSERT_EQ(result.size(), static_cast<size_t>(4));
+  ASSERT_EQ(result.size(), static_cast<size_t>(5));
 }
 
 TEST_F(ManagementFrontendTest, TestAddDuplicateApplication) {
